+++ tests/testSqlite.R
@@ -1,0 +1,3 @@
+library(testthat)
+options(dbms = "sqlite")
+test_check("Strategus")